* Make Olympus decompressor faster (-13%)
* Make Sony LJpeg decompressor faster (-33%)
* Support broken-by-design Blackmagic uncompressed DNG's

* OM System OM-1 Mark II support
* Fujifilm X100VI support
* Sony ILCE-9M3 support
* Phase One P25+ support
* Add Panasonic DC-TZ95D alias
* Remove explicit crop from Fujifilm X-Trans gen-4/5 cameras
* Leica SL3 normalization
* Fix up Panasonic DMC-FZ28 crop modes
* Add unsupported Panasonic placeholders
* Add unsupported Canon placeholders
* Add Fujifilm X100VI placeholder

* upstream/develop: (154 commits)
  Phase One P25+ support
  BitStreamer::getInput(): don't disable fast-path for `DEBUG` builds
  `BitStreamerJPEG::fillCache()`: `prefetch` is no longer needed
  BitStreamer: sink prefetch into `getInput()`
  LJpegDecompressor: change how we deal with predictor
  `BitStreamer::getInput()`: add branch likelyhood annotations
  Bump some fuzz-only image size cut-offs
  `writeLog()`: drop pointless `RAWSPEED_READNONE`, clang complains
  Cr2Decompressor: do not inline `decompressN_X_Y()`'s into `decompress()`
  LJpegDecompressor: do check that jpeg frame has positive area
  TiffParser::parse(): retain already-parsed IFD's on IFD parse failure (Blackmagic oddities)
  CI: macos: bump whole-job timeout
  LJpegDecoder: avoid overflow when computing pixels-per-row of maximal tile
  LJpegDecompressor: drop some more meaningless invariants
  LJpegDecompressor: drop obsolete `cps` variable
  JpegDecompressor: drop arg names from unused arguments
  md5: drop arg name from unused argument
  FiffParser::parseData(): use switch
  OlympusDifferenceDecoder: make definition of `getDiff()` inlne, not decl
  BitStreamerJPEG: make definition of `getStreamPosition()` inlne, not decl
  ...

+++ src/librawspeed/decompressors/SamsungV2Decompressor.cpp
@@ -44,7 +44,7 @@ namespace rawspeed {
 // in contact and Loring von Palleske (Samsung) for pointing to the open-source
 // code of Samsung's DNG converter at http://opensource.samsung.com/
 
-enum struct SamsungV2Decompressor::OptFlags : uint32_t {
+enum struct SamsungV2Decompressor::OptFlags : uint8_t {
   NONE = 0U,       // no flags
   SKIP = 1U << 0U, // Skip checking if we need differences from previous line
   MV = 1U << 1U,   // Simplify motion vector definition
